pyproject.toml
[tool.poetry]
name = "facie"
version = "1.0.0"
description = "Simple friends list service with FastAPI and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.12"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
pillow = "^11.0.0"
openai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.poetry.group.bot.dependencies]
python-telegram-bot = "21.7"
requests = "2.32.3"
python-dotenv = "1.0.1"
